requirements.txt
python-dotenv
langchain-groq
streamlit